+++ src/main/java/com/example/aet/model/item/dto/ItemRequest.java
@@ -7,10 +7,10 @@ import java.time.LocalDateTime;
 
 public record ItemRequest(
         @NotNull(message = "amount may not be null")
-        Long amount,
+        Double amount,
         @NotNull(message = "value may not be null")
         @Min(value = 0, message = "value should be greater than 0")
-        Long value,
+        Double value,
         LocalDateTime date
 ) {
 }
